Add a cache of windows nodes
When creating pods the current logic treats windows and linux nodes the same.
Using a cache of the windows nodes correctly exclude operations that do not
need to be performed.

+++ go-controller/pkg/ovn/ovn.go
@@ -92,6 +92,9 @@ type Controller struct {
 	// to add a egress deny rule.
 	lspEgressDenyCache map[string]int
 
+	// A cache of all the windows nodes on the network
+	windowsNodeCache map[string]bool
+
 	// A mutex for lspIngressDenyCache and lspEgressDenyCache
 	lspMutex *sync.Mutex
 
@@ -336,6 +339,13 @@ func podScheduled(pod *kapi.Pod) bool {
 	return pod.Spec.NodeName != ""
 }
 
+func (oc *Controller) isWindowsNode(nodeName string) bool {
+	if _, exists := oc.windowsNodeCache[nodeName]; exists {
+		return true
+	}
+	return false
+}
+
 // WatchPods starts the watching of Pod resource and calls back the appropriate handler logic
 func (oc *Controller) WatchPods() error {
 	var retryPods sync.Map
@@ -343,6 +353,9 @@ func (oc *Controller) WatchPods() error {
 		AddFunc: func(obj interface{}) {
 			pod := obj.(*kapi.Pod)
 			if !podWantsNetwork(pod) {
+				return
+			}
+			if oc.isWindowsNode(pod.Spec.NodeName) {
 				return
 			}
 
@@ -361,6 +374,9 @@ func (oc *Controller) WatchPods() error {
 			if !podWantsNetwork(pod) {
 				return
 			}
+			if oc.isWindowsNode(pod.Spec.NodeName) {
+				return
+			}
 
 			_, retry := retryPods.Load(pod.UID)
 			if podScheduled(pod) && retry {
@@ -373,6 +389,9 @@ func (oc *Controller) WatchPods() error {
 		},
 		DeleteFunc: func(obj interface{}) {
 			pod := obj.(*kapi.Pod)
+			if oc.isWindowsNode(pod.Spec.NodeName) {
+				return
+			}
 			oc.deleteLogicalPort(pod)
 			retryPods.Delete(pod.UID)
 		},
@@ -506,12 +525,15 @@ func (oc *Controller) WatchNodes() error {
 		AddFunc: func(obj interface{}) {
 			node := obj.(*kapi.Node)
 			logrus.Debugf("Added event for Node %q", node.Name)
+			if node.Status.NodeInfo.OperatingSystem == "windows" {
+				oc.windowsNodeCache[node.Name] = true
+				return
+			}
 			hostSubnet, err := oc.addNode(node)
 			if err != nil {
 				logrus.Errorf("error creating subnet for node %s: %v", node.Name, err)
 				return
 			}
-
 			err = oc.syncNodeManagementPort(node, hostSubnet)
 			if err != nil {
 				macAddressFailed[node.Name] = true
@@ -527,6 +549,11 @@ func (oc *Controller) WatchNodes() error {
 			oldNode := old.(*kapi.Node)
 			node := new.(*kapi.Node)
 			logrus.Debugf("Updated event for Node %q", node.Name)
+			if _, exist := oc.windowsNodeCache[node.Name]; exist && node.Name != oldNode.Name {
+				delete(oc.windowsNodeCache, node.Name)
+				oc.windowsNodeCache[node.Name] = true
+				return
+			}
 			if macAddressFailed[node.Name] || macAddressChanged(oldNode, node) {
 				err := oc.syncNodeManagementPort(node, nil)
 				if err != nil {
@@ -555,6 +582,10 @@ func (oc *Controller) WatchNodes() error {
 			logrus.Debugf("Delete event for Node %q. Removing the node from "+
 				"various caches", node.Name)
 
+			if _, exists := oc.windowsNodeCache[node.Name]; exists {
+				delete(oc.windowsNodeCache, node.Name)
+				return
+			}
 			nodeSubnet, _ := parseNodeHostSubnet(node)
 			err := oc.deleteNode(node.Name, nodeSubnet)
 			if err != nil {
